+++ src/mapdata.h
@@ -237,7 +237,7 @@ struct map_data_common_t {
         * as to which possible object/field/entity in a single square gets drawn and that some symbols
         * are "reserved" such as * and % to do programmatic behavior.
         */
-        std::array<long, SEASONS_PER_YEAR> symbol_;
+        std::array<int, SEASONS_PER_YEAR> symbol_;
 
         int movecost;   // The amount of movement points required to pass this terrain by default.
         int coverage; // The coverage percentage of a furniture piece of terrain. <30 won't cover from sight.
@@ -283,7 +283,7 @@ struct map_data_common_t {
             return ( connect_group != TERCONN_NONE ) && ( connect_group == test_connect_group );
         }
 
-        long symbol() const;
+        int symbol() const;
         nc_color color() const;
 
         const harvest_id &get_harvest() const;
